requirements.txt
anyio==4.7.0
certifi==2024.8.30       
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1     
python-telegram-bot==21.9
requests==2.32.3
sniffio==1.3.1
transmission-rpc==7.0.11 
typing_extensions==4.12.2
urllib3==2.2.3